+++ src/Model/Command/ItemPropertySetup.php
@@ -29,32 +29,32 @@ class ItemPropertySetup extends AbstractCommand
 
     public static function int(string $propertyName): self
     {
-        return new static($propertyName, self::PROPERTY_TYPE_INT);
+        return new static($propertyName, static::PROPERTY_TYPE_INT);
     }
 
     public static function double(string $propertyName): self
     {
-        return new static($propertyName, self::PROPERTY_TYPE_DOUBLE);
+        return new static($propertyName, static::PROPERTY_TYPE_DOUBLE);
     }
 
     public static function string(string $propertyName): self
     {
-        return new static($propertyName, self::PROPERTY_TYPE_STRING);
+        return new static($propertyName, static::PROPERTY_TYPE_STRING);
     }
 
     public static function boolean(string $propertyName): self
     {
-        return new static($propertyName, self::PROPERTY_TYPE_BOOLEAN);
+        return new static($propertyName, static::PROPERTY_TYPE_BOOLEAN);
     }
 
     public static function timestamp(string $propertyName): self
     {
-        return new static($propertyName, self::PROPERTY_TYPE_TIMESTAMP);
+        return new static($propertyName, static::PROPERTY_TYPE_TIMESTAMP);
     }
 
     public static function set(string $propertyName): self
     {
-        return new static($propertyName, self::PROPERTY_TYPE_SET);
+        return new static($propertyName, static::PROPERTY_TYPE_SET);
     }
 
     protected function setPropertyName(string $propertyName): void